requirements.txt
aiogram==2.11.2
aiohttp>=3.6.2
aioredis==1.3.1
python-dotenv~=0.14.0

asyncio~=3.4.3
asyncpg~=0.22.0
requests~=2.24.0